Refactored the code structure of the matrix app in terms of treeview and flat components. Unified the key to hover and search an item in row and column headers.
