+++ src/components/matrix/grid-cell.tsx
@@ -22,26 +22,31 @@ export type GridCellProps = {
 
 const GridCell = ({ columnIndex, rowIndex, data, style }: GridCellProps): JSX.Element => {
   const {
-    hoveredRowIndex,
-    setHoveredRowIndex,
-    hoveredColIndex,
-    setHoveredColIndex,
-    searchedRowIndex,
-    searchedColIndex,
+    searchedRowID,
+    searchedColID,
+    hoveredRowID,
+    setHoveredRowID,
+    hoveredColID,
+    setHoveredColID,
     gridData,
     colorScale,
   } = data;
 
-  const { colors, link, title } = gridData[rowIndex][columnIndex];
+  const { colors, row, column, link, title } = gridData[rowIndex][columnIndex];
+
+  // get id for current row, use it to check hover and search states
+  const { id: rowId } = row;
+  // get id for current column, use it to check hover and search states
+  const { id: columnId } = column;
 
   // className changes based on hovered state
   let gridCellClassName =
-    hoveredRowIndex === rowIndex || hoveredColIndex === columnIndex
+    hoveredRowID === rowId || hoveredColID === columnId
       ? 'grid-cell hovered-cell'
       : 'grid-cell unhovered-cell';
 
   // className changes based on searched state
-  if (searchedRowIndex === rowIndex || searchedColIndex === columnIndex) {
+  if (searchedRowID === rowId || searchedColID === columnId) {
     gridCellClassName += ' searched-cell';
   }
 
@@ -51,21 +56,23 @@ const GridCell = ({ columnIndex, rowIndex, data, style }: GridCellProps): JSX.El
   }
 
   return (
-    <div
-      className={gridCellClassName}
-      style={style}
-      onMouseEnter={() => {
-        setHoveredColIndex(columnIndex);
-        setHoveredRowIndex(rowIndex);
-      }}
-    >
-      {link ? (
-        <a role='button' className='cell-link' href={link} title={title ? title : 'unknown'}>
+    <div>
+      <div
+        className={gridCellClassName}
+        style={style}
+        onMouseEnter={() => {
+          setHoveredRowID(rowId);
+          setHoveredColID(columnId);
+        }}
+      >
+        {link ? (
+          <a role='button' className='cell-link' href={link} title={title ? title : 'unknown'}>
+            <MemoizedColorParts colorScale={colorScale} colors={colors} />
+          </a>
+        ) : (
           <MemoizedColorParts colorScale={colorScale} colors={colors} />
-        </a>
-      ) : (
-        <MemoizedColorParts colorScale={colorScale} colors={colors} />
-      )}
+        )}
+      </div>
     </div>
   );
 };
